Move ClusteringSummary to a separate file.
